+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Enhanced the **unify** command when giving input of a file and not a directory return a clear error message.
 * Added a validation to ensure integrations are not skipped and at least one test playbook is not skipped for each integration or script.
 * Added to the Content Tests support for `context_print_dt`, which queries the incident context and prints the result as a json.
+* Added a validation to ensure that image paths in the README files are valid.
 
 # 1.4.0
 * Enable passing a comma-separated list of paths for the `--input` option of the **lint** command.